EMP-147: Refactored model package to DTO package and added exception handle control to specify HTTP status code
EMP-147: Refactored model package to DTO package and added exception handle control to specify HTTP status code
