+++ equinity-historical-data/src/main/java/uk/gov/justice/laa/crime/equinity/historicaldata/exception/EquinityExceptionResponseHandler.java
@@ -1,5 +1,6 @@
 package uk.gov.justice.laa.crime.equinity.historicaldata.exception;
 
+import jakarta.validation.ConstraintViolationException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
@@ -14,5 +15,13 @@ public class EquinityExceptionResponseHandler {
             ResourceNotFoundException exception
     ) {
         return ResponseEntity.status(HttpStatus.NOT_FOUND).body(exception.getMessage());
+    }
+
+    @ExceptionHandler(ConstraintViolationException.class)
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    public ResponseEntity<String> handleConstraintViolationException(
+            ConstraintViolationException exception
+    ) {
+        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(exception.getMessage());
     }
 }
